# PERF: groupsort_indexer contiguity