Some low-hanging fruit
```
In [1]: import numpy as np
   ...: import pandas._libs.algos as libalgos
   ...:
   ...: np.random.seed(0)
   ...: inds = np.random.randint(0, 100, 500000)
   ...: ngroups = inds.max() + 1

In [2]: %timeit libalgos.groupsort_indexer(inds, ngroups)
```
`1.06 ms ± 11.3 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # this pr`
` 1.2 ms ± 39.1 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # master`